+++ tests/Fixtures/AppBundle/DataTable/Type/RegularPersonTableType.php
@@ -17,7 +17,7 @@ use Omines\DataTablesBundle\Column\DateTimeColumn;
 use Omines\DataTablesBundle\Column\TextColumn;
 use Omines\DataTablesBundle\DataTable;
 use Omines\DataTablesBundle\DataTableEvents;
-use Omines\DataTablesBundle\DataTableTypeInterface;
+use Omines\DataTablesBundle\DataTableType;
 use Omines\DataTablesBundle\Event\DataTablePreResponseEvent;
 
 /**
@@ -25,7 +25,7 @@ use Omines\DataTablesBundle\Event\DataTablePreResponseEvent;
  *
  * @author Niels Keurentjes <niels.keurentjes@omines.com>
  */
-class RegularPersonTableType implements DataTableTypeInterface
+class RegularPersonTableType extends DataTableType
 {
     public function configure(DataTable $dataTable, array $options): void
     {